Fixed bug where run_gridpack.py would try to delete a hepmc file that did not exist (which would throw an error) when using to-step=lhe.

+++ scripts/run_gridpack.py
@@ -158,7 +158,8 @@ if save_hepmc is not None:
     subprocess.check_call(['cp', '%s/events_%i.hepmc.gz' % (tmpdir, seed), '%s/events_%i.hepmc.gz' % (save_hepmc, seed)])
     os.remove('%s/events_%i.hepmc.gz' % (tmpdir, seed))
 else:
-    os.remove('%s/events_%i.hepmc' % (tmpdir, seed))
+    if not args.to_step == 'lhe': #if stopped at lhe, there would be no hepmc to delete
+        os.remove('%s/events_%i.hepmc' % (tmpdir, seed))
 
 if not args.no_cleanup and load_hepmc is None:
     subprocess.check_call(['rm', '-rf', gridpack_dir])
